backend/requirements.txt
# Core Backend
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# AI/ML - RAG Pipeline
transformers>=4.35.0
sentence-transformers>=2.2.2
chromadb>=0.4.18
huggingface-hub>=0.19.4
torch>=2.0.0

# Data Processing
pandas>=2.1.0
numpy>=1.26.0
openpyxl>=3.1.0
requests>=2.31.0

# OCR and PDF Parsing
pdfplumber>=0.10.0
# paddleocr>=2.7.0  # Optional - can cause conflicts, install separately if needed
# paddlepaddle>=2.5.0  # Optional
Pillow>=10.0.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0

# Visualization
matplotlib>=3.8.0
plotly>=5.18.0
geopandas>=0.14.0
kaleido>=0.2.0

# Task Orchestration
# prefect>=2.14.0  # Optional - install separately if needed

# Utilities
python-dotenv>=1.0.0
python-multipart>=0.0.6
httpx>=0.25.0
aiofiles>=23.0.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.1.0
